Add placeholder to thumbnails in case thumbnail is missing
